refactor(useCases/deleteUser): :recycle: add validation

+++ src/domain/useCases/users/deleteUser.ts
@@ -4,6 +4,12 @@ export class DeleteUser {
     constructor(private userRepository: UserRepository) {}
 
     async execute(id: string): Promise<void> {
+        const user = await this.userRepository.findOne(id);
+
+        if (!user) {
+            throw new Error(`User with id ${id} not found`);
+        }
+
         await this.userRepository.remove(id);
     }
 }
